Refactor python code to reflect SDK changes
