+++ hyveos_bridge/kv.py
@@ -1,41 +1,41 @@
-from hyveos_msgs.srv import DHTGetRecord, DHTPutRecord
-from hyveos_sdk import DHTService
+from hyveos_msgs.srv import GetKVRecord, PutKVRecord
+from hyveos_sdk import KVService
 from rclpy.impl.rcutils_logger import RcutilsLogger
 
 from .bridge import Bridge, BridgeClient, service_callback
 
 
-class DHTClient(BridgeClient):
+class KVClient(BridgeClient):
     logger: RcutilsLogger
-    dht: DHTService
+    kv: KVService
 
     def __init__(self, node: Bridge):
         def namespaced(name: str) -> str:
-            return f'{node.get_name()}/dht/{name}'
+            return f'{node.get_name()}/kv/{name}'
 
         self.get_record_service = node.create_service(
-            DHTGetRecord,
+            GetKVRecord,
             namespaced('get_record'),
             self._get_record_callback
         )
         self.put_record_service = node.create_service(
-            DHTPutRecord,
+            PutKVRecord,
             namespaced('put_record'),
             self._put_record_callback
         )
 
         self.logger = node.get_logger()
-        self.dht = node.connection.get_dht_service()
+        self.kv = node.connection.get_kv_service()
 
     @service_callback
     async def _get_record_callback(
         self,
-        request: DHTGetRecord.Request,
-        response: DHTGetRecord.Response
+        request: GetKVRecord.Request,
+        response: GetKVRecord.Response
     ):
         self.logger.info(f'Getting record in topic {request.topic} with key {request.key}')
 
-        record = await self.dht.get_record(request.topic, request.key)
+        record = await self.kv.get_record(request.topic, request.key)
 
         if record.data is None:
             response.success = False
@@ -49,12 +49,12 @@ class DHTClient(BridgeClient):
     @service_callback
     async def _put_record_callback(
         self,
-        request: DHTPutRecord.Request,
-        response: DHTPutRecord.Response
+        request: PutKVRecord.Request,
+        response: PutKVRecord.Response
     ):
         self.logger.info(f'Putting record in topic {request.topic} with key {request.key}')
 
-        await self.dht.put_record(request.topic, request.key, request.value)
+        await self.kv.put_record(request.topic, request.key, request.value)
 
         response.success = True
         return response
